Session serialization to zipped yaml/tsv
Related to #24. Serialize session data to a zipped yaml file and tsv files instead of pickling, so the data is portable across different versions of Python and other dependencies.

Currently, this is still too slow. If we can reduce the amount of data without losing much, this could be become viable.